OrangeJuice.Server.Core/Data/AwsFoodRepository.cs
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using System.Xml.Linq;

using OrangeJuice.Server.Filters;
using OrangeJuice.Server.Services;

namespace OrangeJuice.Server.Data
{
	public sealed class AwsFoodRepository : IFoodRepository
	{
		private readonly IAwsProvider _provider;
		private readonly IFoodDescriptionFactory _foodDescriptionFactory;
		private readonly IFilter<FoodDescription> _foodDescriptionFilter;

		public AwsFoodRepository(IAwsProvider provider, IFoodDescriptionFactory foodDescriptionFactory, IFilter<FoodDescription> foodDescriptionFilter)
		{
			_provider = provider;
			_foodDescriptionFactory = foodDescriptionFactory;
			_foodDescriptionFilter = foodDescriptionFilter;
		}

		public async Task<ICollection<FoodDescription>> Search(string title)
		{
			ICollection<XElement> items = await _provider.SearchItems(title);
			return items.Select(i => _foodDescriptionFactory.Create(i))
						.Where(_foodDescriptionFilter.Filter)
						.ToArray();
		}
	}
}